stack/streamlit/requirements.txt
streamlit==1.29.0
pandas==2.1.4
numpy==1.25.2
plotly==5.17.0
requests==2.31.0
psycopg2-binary==2.9.9
pymongo==4.6.0
kafka-python==2.0.2
minio==7.2.0
